feat: add constructor and method that take a bundle or omitting locale
