+++ src/org/openide/awt/AbstractMnemonicsAction.java
@@ -2,11 +2,37 @@ package org.openide.awt;
 
 import javax.swing.*;
 import java.util.Locale;
+import java.util.ResourceBundle;
 
-public abstract class AbstractMnemonicsAction extends AbstractAction implements MnemonicsAction {
+/**
+ * @author Hiroshi Miura
+ */
+public abstract class AbstractMnemonicsAction extends AbstractAction {
+    public AbstractMnemonicsAction() {
+        super();
+    }
+
+    public AbstractMnemonicsAction(String key, ResourceBundle bundle) {
+        super();
+        setText(key, bundle);
+    }
+
     public AbstractMnemonicsAction(String text, Locale locale) {
         super();
         setText(text, locale);
+    }
+
+    public AbstractMnemonicsAction(String text) {
+        super();
+        setText(text);
+    }
+
+    public void setText(String key, ResourceBundle bundle) {
+        setText(bundle.getString(key), bundle.getLocale());
+    }
+
+    public void setText(String text) {
+        setText(text, new Locale("en"));
     }
 
     public void setText(String text, Locale locale) {
